pyproject.toml
[project]
name = "rec-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crawl4ai>=0.7.6",
    "faiss-cpu>=1.12.0",
    "flask>=3.1.2",
    "langchain-text-splitters>=1.0.0",
    "llama-index>=0.14.7",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-embeddings-huggingface-api>=0.4.1",
    "llama-index-llms-gemini>=0.6.1",
    "llama-index-llms-google-genai>=0.7.3",
    "llama-index-vector-stores-faiss>=0.5.1",
    "streamlit>=1.51.0",
]
